u8 returns cost nothing in the callee and an andi in every caller: 384
func_80059FAC matched immediately with the arm-scoped base pointer -- the
lever from func_8002CCE4, third function it has worked on -- and then broke
when I gave its two callees their real prototypes. Both are declared
`u8 f(void)` and return a u8 global; gcc will not trust a callee to have
extended its return value, so the caller emits `andi $r,$r,255` after each
call. Retail masks neither.

Widening both definitions to `s32` leaves them byte-identical -- the value
comes from an lbu either way -- and lets the caller match. Which means the
narrow return type was never doing anything except waiting to cost an
instruction in the first function that called it, while the callee itself
stayed green and made it look like the caller's fault.

+++ src/func_8005F174.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8005F174(void) {
+s32 func_8005F174(void) {
     return D_8009B07B;
 }
