+++ src/func_8005F18C.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8005F18C(void) {
+s32 func_8005F18C(void) {
     return D_8009B07C;
 }
